change:  - cleanup old notification class & reducer.  - removed pager.modules.scss & added style.ts.
