--- packages/boba/gateway/src/components/pager/Pager.module.scss
@@ -1,55 +1,0 @@
-@import 'index.scss';
-
-.Pager {
-  // background: $gray6;
-  display: flex;
-  flex-direction: row;
-  min-height: 25px;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 6px;
-  padding: 10px;
-
-  .number {
-    // font-size: 0.8em;
-    // padding-top: 5px;
-  }
-
-  .numberLeft {
-    // font-size: 1.0em;
-    margin-left: 10px;
-  }
-
-  .numberRight {
-    display: flex;
-    flex-direction: row;
-    margin-right: 10px;
-    align-items: center;
-  }
-
-  .box {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    /*margin-left: 3px;*/
-    transition: all 200ms ease-in-out;
-    cursor: pointer;
-    // background-color: #506DFA;//$green;
-    border: 2px solid #506dfa;
-    padding: 2px;
-    border-radius: 8px; /// 100%;
-    width: 25px;
-    height: 25px;
-    margin: 0px 5px;
-
-    &:hover {
-      background-color: #506dfa; //$green;
-    }
-  }
-
-  .disabled {
-    pointer-events: none;
-    cursor: initial;
-    opacity: 0.5;
-  }
-}
