+++ packages/boba/gateway/src/components/select/Select.module.scss
@@ -1,5 +1,3 @@
-/* @import 'index.scss'; */
-
 .Select {
   display: flex;
   flex-direction: column;
